requirements.txt
tensorboard==2.10.0
tensorflow==2.7.0
tensorflow-addons==0.17.1
opencv_python==4.2.0.34
matplotlib==3.2.0
pandas==1.2.3
tqdm==4.40.2
numpy==1.19.5
pyyaml==5.4.1
scikit-learn==0.24.1
scikit-optimize==0.8.1
dill==0.3.3
pillow==8.3.2
onnx==1.11.0
onnx-tf==1.10.0
